requirements.txt
streamlit
fpdf2
jinja2
markdown2
pdfkit
wkhtmltopdf
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.12
langchain-community>=0.0.15
langchain-text-splitters>=0.0.1
pypdf
chromadb
python-json-logger
faiss-cpu
pdf2image
pillow
pytesseract
pymupdf